docs: Change in the bikeshare python script
Changes were implemented in the python script where we added comments
above the month, day, and city data which are placed in lists and
dictionaries. The comments help explain to the reader of the
importance of the dictionaries and lists in the code.

Commit for doc change in bikeshare python script